dashboard-hospedaje/requirements.txt
gspread
google-auth
pandas
plotly
reportlab
python-dotenv
numpy
matplotlib
